Mostly fix LGTM warnings.

+++ src/main/java/org/apache/datasketches/characterization/hll/DruidHllAccuracyProfile.java
@@ -42,11 +42,11 @@ public class DruidHllAccuracyProfile extends BaseAccuracyProfile {
   public void doTrial() {
     final int qArrLen = qArr.length;
     sketch = HyperLogLogCollector.makeLatestCollector();
-    int lastUniques = 0;
+    long lastUniques = 0;
     for (int i = 0; i < qArrLen; i++) {
       final AccuracyStats q = qArr[i];
-      final double delta = q.trueValue - lastUniques;
-      for (int u = 0; u < delta; u++) {
+      final long delta = (long)(q.trueValue - lastUniques);
+      for (long u = 0; u < delta; u++) {
         if (useString) {
           final String vstr = Long.toHexString(++vIn);
           sketch.add(hash.hash(vstr));
