+++ src/main/java/org/apache/datasketches/characterization/theta/ThetaAccuracyProfile.java
@@ -63,11 +63,11 @@ public class ThetaAccuracyProfile extends BaseAccuracyProfile {
   public void doTrial() {
     final int qArrLen = qArr.length;
     sketch.reset(); //reuse the same sketch
-    int lastUniques = 0;
+    long lastUniques = 0;
     for (int i = 0; i < qArrLen; i++) {
       final AccuracyStats q = qArr[i];
-      final double delta = q.trueValue - lastUniques;
-      for (int u = 0; u < delta; u++) {
+      final long delta = (long)(q.trueValue - lastUniques);
+      for (long u = 0; u < delta; u++) {
         sketch.update(++vIn);
       }
       lastUniques += delta;
